+++ python/mlc_llm/model/gemma3/gemma3_loader.py
@@ -13,9 +13,7 @@ from mlc_llm.quantization import Quantization
 from .gemma3_model import Gemma3Config, Gemma3ForCausalLM
 
 
-def huggingface(
-    model_config: Gemma3Config, quantization: Quantization
-) -> ExternMapping:
+def huggingface(model_config: Gemma3Config, quantization: Quantization) -> ExternMapping:
     """Returns a parameter mapping that maps from the names of MLC LLM parameters to
     the names of HuggingFace PyTorch parameters.
 
@@ -57,9 +55,7 @@ def huggingface(
                 f"{hf_prefix + mlp}.up_proj.weight",
             ],
             functools.partial(
-                lambda gate, up, dtype: np.concatenate([gate, up], axis=0).astype(
-                    dtype
-                ),
+                lambda gate, up, dtype: np.concatenate([gate, up], axis=0).astype(dtype),
                 dtype=mlc_param.dtype,
             ),
         )
