+++ python/mlc_llm/model/stable_lm/stablelm_loader.py
@@ -13,9 +13,7 @@ from mlc_llm.quantization import Quantization
 from .stablelm_model import StableLmConfig, StableLmForCausalLM
 
 
-def huggingface(
-    model_config: StableLmConfig, quantization: Quantization
-) -> ExternMapping:
+def huggingface(model_config: StableLmConfig, quantization: Quantization) -> ExternMapping:
     """Returns a parameter mapping that maps from the names of MLC LLM parameters to
     the names of HuggingFace PyTorch parameters.
 
@@ -73,9 +71,7 @@ def huggingface(
                     f"{attn}.v_proj.bias",
                 ],
                 functools.partial(
-                    lambda q, k, v, dtype: np.concatenate([q, k, v], axis=0).astype(
-                        dtype
-                    ),
+                    lambda q, k, v, dtype: np.concatenate([q, k, v], axis=0).astype(dtype),
                     dtype=mlc_param.dtype,
                 ),
             )
@@ -90,9 +86,7 @@ def huggingface(
                 f"{mlp}.up_proj.weight",
             ],
             functools.partial(
-                lambda gate, up, dtype: np.concatenate([gate, up], axis=0).astype(
-                    dtype
-                ),
+                lambda gate, up, dtype: np.concatenate([gate, up], axis=0).astype(dtype),
                 dtype=mlc_param.dtype,
             ),
         )
