+++ python/mlc_llm/model/starcoder2/starcoder2_loader.py
@@ -13,9 +13,7 @@ from mlc_llm.quantization import Quantization
 from .starcoder2_model import Starcoder2Config, Starcoder2ForCausalLM
 
 
-def huggingface(
-    model_config: Starcoder2Config, quantization: Quantization
-) -> ExternMapping:
+def huggingface(model_config: Starcoder2Config, quantization: Quantization) -> ExternMapping:
     """Returns a parameter mapping that maps from the names of MLC LLM parameters to
     the names of HuggingFace PyTorch parameters.
 
@@ -82,9 +80,7 @@ def huggingface(
                     f"{attn}.v_proj.bias",
                 ],
                 functools.partial(
-                    lambda q, k, v, dtype: np.concatenate([q, k, v], axis=0).astype(
-                        dtype
-                    ),
+                    lambda q, k, v, dtype: np.concatenate([q, k, v], axis=0).astype(dtype),
                     dtype=mlc_param.dtype,
                 ),
             )
